[nomaster] inline importPrivateWithinFromJavaFlags into SymbolTable
This reworks 02ed5fb so that we don't change JavaUniverse's super classes.

This is necessary to maintain binary compatibility with 2.10.0.
