+++ bincompat-forward.whitelist.conf
@@ -85,10 +85,10 @@ filter {
         #     matchName="scala.reflect.runtime.SymbolLoaders.initClassModule"
         #     problemName=MissingMethodProblem
         # },
-        {
-            matchName="scala.reflect.runtime.JavaUniverse"
-            problemName=MissingTypesProblem
-        },
+        # {
+        #     matchName="scala.reflect.runtime.JavaUniverse"
+        #     problemName=MissingTypesProblem
+        # },
         # {
         #     matchName="scala.reflect.runtime.JavaUniverse.initClassAndModule"
         #     problemName=MissingMethodProblem
